Adding the fully parametric model
Now it has a more tensorflow spirit. We profit from the keras.Layer and keras.Model to build the PSF layers and Models.
Now the model is able to compile and there is no need to run in eager mode. The problem was that the when compiled is it needed to use the tf.shape() function and not the tensor.shape. As the first one is calculated on runtime and the second one in compilation time!
There is a possible issue that has to be investigated. If the current_opd vairiable from the class is being shared between the map_fn() call or not. It shouldn't.